+++ FluentPassFinderContracts/Ipc/PipeRequestTypes.cs
@@ -1,19 +1,5 @@
 namespace FluentPassFinder.Contracts.Public.Ipc
 {
-    /// <summary>Base class for all pipe request messages.</summary>
-    public class PipeRequest
-    {
-        public string Id { get; set; }
-        public string Type { get; set; }
-    }
-
-    /// <summary>Base class for all pipe response messages.</summary>
-    public class PipeResponse
-    {
-        public string Id { get; set; }
-        public bool Success { get; set; }
-        public string Error { get; set; }
-    }
 
     public static class PipeRequestTypes
     {
